rename colour to color
